# contracts/product/v1/openapi.yaml
openapi: 3.1.2
...
info:
  title: Cyrene Reactor Product API
  version: 1.0.0
  license:
    name: Apache-2.0
paths:
  /api/v1/deployments:
    post:
      operationId: createDeployment
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeploymentRequest'
      responses:
        '201':
          description: Reconciled Deployment
          content:
            application/json:
              schema:
                $ref: ./deployment.schema.json
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
    get:
      summary: List Deployments
      operationId: list_deployments_api_v1_deployments_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Deployment'
                title: Response List Deployments Api V1 Deployments Get
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /api/v1/deployments/{deploymentId}:
    get:
      operationId: getDeployment
      parameters:
      - $ref: '#/components/parameters/DeploymentId'
      responses:
        '200':
          description: Deployment after endpoint identity reconciliation
          content:
            application/json:
              schema:
                $ref: ./deployment.schema.json
        '404':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /api/v1/deployments/{deploymentId}/events:
    get:
      operationId: getDeploymentEvents
      parameters:
      - $ref: '#/components/parameters/DeploymentId'
      responses:
        '200':
          description: Historical list of phase events for a deployment
          content:
            application/json:
              schema:
                type: object
                required: [deploymentId, events]
                properties:
                  deploymentId: {type: string, format: uuid}
                  events:
                    type: array
                    items:
                      type: object
                      required: [sequence, phase, message, occurredAt]
                      properties:
                        sequence: {type: integer, minimum: 1}
                        phase: {type: string}
                        message: {type: string}
                        occurredAt: {type: string, format: date-time}
                        failureCode: {type: string, nullable: true}
        '404':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /api/v1/deployments/{deploymentId}/actions/stop:
    post:
      operationId: stopDeployment
      parameters:
      - $ref: '#/components/parameters/DeploymentId'
      responses:
        '200':
          description: Stopped Deployment
          content:
            application/json:
              schema:
                $ref: ./deployment.schema.json
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /api/v1/endpoints/{endpointId}:
    get:
      operationId: getEndpoint
      parameters:
      - name: endpointId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Endpoint resource
          content:
            application/json:
              schema:
                $ref: ./endpoint.schema.json
        '404':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /api/v1/deployments/{deploymentId}/export:
    get:
      summary: Export Deployment
      operationId: export_deployment_api_v1_deployments__deploymentId__export_get
      parameters:
      - name: deploymentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Deploymentid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Export Deployment Api V1 Deployments  Deploymentid  Export
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /api/v1/deployments/{deploymentId}/actions/restart:
    post:
      summary: Restart
      operationId: restart_api_v1_deployments__deploymentId__actions_restart_post
      parameters:
      - name: deploymentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Deploymentid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestartRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /api/v1/serving-bindings:
    get:
      summary: Bindings
      operationId: bindings_api_v1_serving_bindings_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                type: array
                title: Response Bindings Api V1 Serving Bindings Get
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /api/v1/serving-bindings/{binding_id}/node:
    get:
      summary: Node
      operationId: node_api_v1_serving_bindings__binding_id__node_get
      parameters:
      - name: binding_id
        in: path
        required: true
        schema:
          type: string
          title: Binding Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Node Api V1 Serving Bindings  Binding Id  Node Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /api/v1/model-imports:
    post:
      summary: Create Model Import
      operationId: create_model_import_api_v1_model_imports_post
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateModelImportRequest'
      responses:
        '201':
          description: Validated ModelImport
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelImport'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
    get:
      summary: List Model Imports
      operationId: list_model_imports_api_v1_model_imports_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModelImport'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /api/v1/model-imports/{importId}:
    get:
      summary: Get Model Import
      operationId: get_model_import_api_v1_model_imports__importId__get
      parameters:
      - name: importId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: ModelImport resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelImport'
        '404':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /api/v1/exchange-receivers:
    get:
      summary: Receivers
      operationId: receivers_api_v1_exchange_receivers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                type: array
                title: Response Receivers Api V1 Exchange Receivers Get
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /api/v1/endpoints/{endpointId}/actions/send-to-exchange:
    post:
      summary: Send To Exchange
      operationId: send_to_exchange_api_v1_endpoints__endpointId__actions_send_to_exchange_post
      parameters:
      - name: endpointId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Endpointid
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendToExchangeRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Send To Exchange Api V1 Endpoints  Endpointid  Actions
                  Send To Exchange Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /api/v1/deployments/{deploymentId}/model-export:
    get:
      summary: Export Model
      operationId: export_model_api_v1_deployments__deploymentId__model_export_get
      parameters:
      - name: deploymentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Deploymentid
      responses:
        '200':
          description: Complete standard Hugging Face model directory, including upstream
            license
          content:
            application/x-tar:
              schema:
                type: string
                format: binary
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 200
    DeploymentId:
      name: deploymentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content:
        application/problem+json:
          schema:
            $ref: ./generated/common/problem-details.schema.json
  schemas:
    CreateDeploymentRequest:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        modelArtifact:
          $ref: '#/components/schemas/ArtifactRef'
        composition:
          type: string
          enum:
          - FULL_MODEL
          - BASE_PLUS_LORA
          default: FULL_MODEL
        modelVersion:
          $ref: '#/components/schemas/ModelVersion'
        servingBindingId:
          type: string
          maxLength: 200
          minLength: 1
          title: Servingbindingid
        nodeRef:
          anyOf:
          - $ref: '#/components/schemas/NodeRef'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - name
      - servingBindingId
      title: CreateDeploymentRequest
      description: Create-Deployment command. | 创建 Deployment 请求。
      allOf:
      - if:
          required:
          - composition
          properties:
            composition:
              const: BASE_PLUS_LORA
        then:
          required:
          - modelVersion
          properties:
            modelVersion:
              required:
              - composition
              properties:
                composition:
                  const: BASE_PLUS_LORA
      - if:
          required:
          - composition
          - modelVersion
          properties:
            composition:
              const: FULL_MODEL
        then:
          properties:
            modelVersion:
              required:
              - composition
              properties:
                composition:
                  const: FULL_MODEL
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    CreateModelImportRequest:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        servingBindingId:
          type: string
          maxLength: 200
          minLength: 1
          title: Servingbindingid
        source:
          $ref: '#/components/schemas/ModelImportSource'
        credentialRef:
          anyOf:
          - type: string
            maxLength: 300
            minLength: 1
          - type: 'null'
        trustRemoteCode:
          type: boolean
          title: Trustremotecode
          default: false
      additionalProperties: false
      type: object
      required:
      - name
      - servingBindingId
      - source
      title: CreateModelImportRequest
      description: Create-ModelImport command. | 创建模型导入请求。
...
    NodeRef:
      properties:
        nodeId:
          type: string
          maxLength: 200
          minLength: 1
          title: Nodeid
        nodeEpoch:
          type: integer
          exclusiveMinimum: 0.0
          title: Nodeepoch
      additionalProperties: false
      type: object
      required:
      - nodeId
      - nodeEpoch
      title: NodeRef
      description: Projection of the canonical Node identity and epoch. | 节点身份与纪元投影。
...
    RestartRequest:
      properties:
        nodeRef:
          $ref: '#/components/schemas/NodeRef'
        resourceVersion:
          type: integer
          exclusiveMinimum: 0.0
          title: Resourceversion
      additionalProperties: false
      type: object
      required:
      - nodeRef
      - resourceVersion
      title: RestartRequest
      description: Restart requires refreshed identity and current Product version.
        | 重启确认。
...
  securitySchemes:
    ControlBearer:
      type: http
      scheme: bearer
security:
- ControlBearer: []